+++ damc/damc_common/BiquadFilter.cpp
@@ -3,11 +3,11 @@
 #include <math.h>
 #include <string.h>
 
-void BiquadFilter::init(const double a_coefs[], const double b_coefs[]) {
+void BiquadFilter::init(const float a_coefs[], const float b_coefs[]) {
 	update(a_coefs, b_coefs);
 }
 
-void BiquadFilter::update(const double a_coefs[], const double b_coefs[]) {
+void BiquadFilter::update(const float a_coefs[], const float b_coefs[]) {
 	this->b_coefs[0] = b_coefs[0] / a_coefs[0];
 	this->b_coefs[1] = b_coefs[1] / a_coefs[0];
 	this->b_coefs[2] = b_coefs[2] / a_coefs[0];
@@ -15,26 +15,26 @@ void BiquadFilter::update(const double a_coefs[], const double b_coefs[]) {
 	this->a_coefs[1] = a_coefs[2] / a_coefs[0];
 }
 
-double BiquadFilter::put(double input) {
-	const double* const a = a_coefs;
-	const double* const b = b_coefs;
+float BiquadFilter::put(float input) {
+	const float* const a = a_coefs;
+	const float* const b = b_coefs;
 
-	double y = b[0] * input + s1 - 0.5;
+	float y = b[0] * input + s1 - 0.5;
 	s1 = s2 + b[1] * input - a[0] * y;
 	s2 = b[2] * input - a[1] * y + 0.5;
 
 	return y;
 }
 
-std::complex<double> BiquadFilter::getResponse(double f0, double fs) {
-	const double* const a = a_coefs;
-	const double* const b = b_coefs;
+std::complex<float> BiquadFilter::getResponse(float f0, float fs) {
+	const float* const a = a_coefs;
+	const float* const b = b_coefs;
 
-	double w = 2 * M_PI * f0 / fs;
-	std::complex<double> z_1 = std::exp(std::complex<double>(0, -1 * w));
-	std::complex<double> z_2 = std::exp(std::complex<double>(0, -2 * w));
+	float w = 2 * M_PI * f0 / fs;
+	std::complex<float> z_1 = std::exp(std::complex<float>(0, -1 * w));
+	std::complex<float> z_2 = std::exp(std::complex<float>(0, -2 * w));
 
-	return (b[0] + b[1] * z_1 + b[2] * z_2) / (std::complex<double>(1) + a[0] * z_1 + a[1] * z_2);
+	return (b[0] + b[1] * z_1 + b[2] * z_2) / (std::complex<float>(1) + a[0] * z_1 + a[1] * z_2);
 }
 
 void BiquadFilter::computeFilter(bool enabled,
@@ -43,8 +43,8 @@ void BiquadFilter::computeFilter(bool enabled,
                                  float fs,
                                  float gain,
                                  float Q,
-                                 double a_coefs[3],
-                                 double b_coefs[3]) {
+                                 float a_coefs[3],
+                                 float b_coefs[3]) {
 	if(!enabled || Q == 0 || fs == 0) {
 		b_coefs[0] = 1;
 		b_coefs[1] = 0;
@@ -53,9 +53,9 @@ void BiquadFilter::computeFilter(bool enabled,
 		a_coefs[1] = 0;
 		a_coefs[2] = 0;
 	} else {
-		double A = pow(10, gain / 40);
-		double w0 = 2 * M_PI * f0 / fs;
-		double alpha = sin(w0) / (2 * Q);
+		float A = pow(10, gain / 40);
+		float w0 = 2 * M_PI * f0 / fs;
+		float alpha = sin(w0) / (2 * Q);
 
 		switch(filterType) {
 			default:
@@ -144,8 +144,8 @@ void BiquadFilter::computeFilter(bool enabled,
 }
 
 void BiquadFilter::computeFilter(bool enabled, FilterType filterType, float f0, float fs, float gain, float Q) {
-	double a_coefs[3];
-	double b_coefs[3];
+	float a_coefs[3];
+	float b_coefs[3];
 
 	BiquadFilter::computeFilter(enabled, filterType, f0, fs, gain, Q, a_coefs, b_coefs);
 	update(a_coefs, b_coefs);
