+++ damc/damc_common/BiquadFilter.h
@@ -18,9 +18,9 @@ enum class FilterType {
 
 class BiquadFilter {
 public:
-	void init(const double a_coefs[3], const double b_coefs[3]);
-	void update(const double a_coefs[3], const double b_coefs[3]);
-	double put(double input);
+	void init(const float a_coefs[3], const float b_coefs[3]);
+	void update(const float a_coefs[3], const float b_coefs[3]);
+	float put(float input);
 
 	static void computeFilter(bool enabled,
 	                          FilterType filterType,
@@ -28,15 +28,15 @@ public:
 	                          float fs,
 	                          float gain,
 	                          float Q,
-	                          double a_coefs[3],
-	                          double b_coefs[3]);
+	                          float a_coefs[3],
+	                          float b_coefs[3]);
 	void computeFilter(bool enabled, FilterType filterType, float f0, float fs, float gain, float Q);
-	std::complex<double> getResponse(double f0, double fs);
+	std::complex<float> getResponse(float f0, float fs);
 
 private:
 	// Use offset of 0.5 to avoid denormals
-	double s1 = 0.5;
-	double s2 = 0.5;
-	double b_coefs[3] = {};
-	double a_coefs[2] = {};
+	float s1 = 0.5;
+	float s2 = 0.5;
+	float b_coefs[3] = {};
+	float a_coefs[2] = {};
 };
